refactor(userspace-dp): extract SnapshotIntegrityError into a submodule
Addresses #4421 (policy.rs modularity backlog) with a bounded, pure
code-motion first increment: relocate the `SnapshotIntegrityError` enum
plus its `Display` and `Error` impls (~888 LOC, the #1606 snapshot
integrity error group) out of the 4483-LOC `userspace-dp/src/policy.rs`
into a new `#[path]`-included sibling submodule
`userspace-dp/src/policy_snapshot_error.rs`.

The enum group is fully self-contained — variants use only primitive /
prelude types (String, i32, u16, u32, u8), the `Display` impl is entirely
`write!(f, ...)` over `Self::` variants, and the `Error` impl is empty, so
the new file needs no imports. It is re-exported with
`pub(crate) use snapshot_error::SnapshotIntegrityError;` so all ~20
crate-wide call sites keep the existing `crate::policy::SnapshotIntegrityError`
path, and `policy_tests.rs` (attached via `use super::*`) is unchanged.
This mirrors the `#[path = "policy_tests.rs"]` idiom already used in the
same file.

This is verified pure code-motion, not a behavior change:
- the moved block is byte-identical to master `policy.rs` lines 14-901
  (empty diff);
- `policy.rs`'s retained content is byte-identical (empty diff) with only
  a 3-line `#[path]`/`mod`/`use` stub inserted;
- `policy.rs` shrinks 4483 -> 3598 LOC.

The moved block is left byte-identical (not rustfmt'd) to preserve the
code-motion guarantee; only the new stub + module-doc lines are new.

Validation: `cargo build` green; `cargo test` 3738 passed (245 policy
tests + all 5 snapshot-integrity-error tests green). The single
`native_gre_decap_checksum_present_yields_inner_packet` failure in the
parallel run is a pre-existing shared-counter test-isolation flake — it
touches no SnapshotIntegrityError code and passes in isolation.

The README source-map table row for `policy.rs` is unchanged: it lists
top-level feature modules, not their `#[path]` submodules (`policy_tests.rs`
is likewise omitted), and the module contract is unchanged. The deeper
policy.rs decomposition (parse / evaluate / applications splits) is a
function-restructuring task that needs /triple-review; this ships only the
clean, self-contained down-payment.
